crisismap_ai/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pymongo>=4.0.0
transformers>=4.40.0
torch>=2.0.0
sentence-transformers>=2.2.0
python-dotenv>=0.19.0
pydantic>=1.8.0
tqdm>=4.62.0
numpy>=1.20.0
beautifulsoup4>=4.10.0
requests>=2.25.0
jinja2>=3.0.0
python-multipart>=0.0.5
typing-extensions>=4.0.0
dnspython>=2.0.0
colorama>=0.4.4 